[remake?] sort of reorgniaze.
